backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
python-jose[cryptography]>=3.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.2.0

# LangChain ecosystem
langchain>=0.2.0
langchain-text-splitters>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.0.3

# Vector store
chromadb>=0.5.0

# Embeddings
sentence-transformers>=3.0.0

# Document loaders
pypdf>=4.2.0
pandas>=2.2.0
openpyxl>=3.1.0
unstructured>=0.14.0
